+++ src/api/mod.rs
@@ -1076,6 +1076,7 @@ mod fetch_integration {
             auth: Default::default(),
             oauth: Default::default(),
             create: Default::default(),
+            hooks: Default::default(),
             view_jql: Config::build_view_jql(&Default::default()),
         }
     }